requirements.txt
Django==1.8.2
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-debug-toolbar==1.2.2
django-extensions==1.5.5
django-filter==0.10.0
django-guardian==1.1.1
django-pylibmc==0.5.0
django-storages==1.1.6
django-stormpath==1.0.3
djangorestframework==3.1.3
httpie
ipython==3.0.0
Markdown
newrelic==2.36.0.30
numpy==1.9.2
psycopg2==2.5.1
pylibmc==1.3.0
redis==2.9.1
requests==2.4.3
six==1.8.0
virtualenv
waitress==0.8.9
Werkzeug
wsgiref==0.1.2
